Feature/details screen v0.1

+++ app/src/main/java/com/onopry/movieapp/presentation/lists/adapters/MovieAdapter.kt
@@ -6,11 +6,11 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.onopry.movieapp.R
-import com.onopry.movieapp.data.models.movie.MoviePreviewItemResponseBody
+import com.onopry.movieapp.data.models.movie.preview.MoviePreviewItemResponseBody
 import com.onopry.movieapp.databinding.ItemMovieListBinding
 import com.onopry.movieapp.presentation.lists.diffutills.MovieDiffUtillCallback
 
-typealias OnRecyclerViewItemClickListener = (movieId: Int) -> Unit
+typealias OnRecyclerViewItemClickListener = (movieId: Long) -> Unit
 
 class MovieAdapter(
     private val clickListener: OnRecyclerViewItemClickListener
